4 Jan 2024, 13:30

+++ swapAlternatives.cpp
@@ -23,7 +23,7 @@ int main(){
     //     cout<<arr[i]<<" ";
     // }
 
-    int i=1;
+    short int i=1;
     while(i){
         cout<<i<<endl;
         i++;
